+++ src/posthog-surveys-types.ts
@@ -53,11 +53,11 @@ export type SurveyQuestionDescriptionContentType = 'html' | 'text'
 
 interface SurveyQuestionBase {
     question: string
+    id?: string
     description?: string | null
     descriptionContentType?: SurveyQuestionDescriptionContentType
     optional?: boolean
     buttonText?: string
-    originalQuestionIndex: number
     branching?: NextQuestionBranching | EndBranching | ResponseBasedBranching | SpecificQuestionBranching
 }